// visor-angular-main/src/app/core/models/idocumento.ts
export interface IDocumento {


    idDocumento: number;
	tipodoc: string;
	documento: string;
	razonsocial: string;
	apepat: string;
	apemat: string;
	nombres: string;
	documentoname: string;
	path: string;
	documentofisico: string;
	tnumpages: number;
	tnumrows: number;
	barcofscan: string;
  
  }